Started dividing code up in to different files, need to do some more code clean up! Still working on CLI
