+++ berrfuzz.go
@@ -10,7 +10,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
-	"math/bits"
 	mrand "math/rand"
 	"os"
 	"os/exec"
@@ -97,74 +96,6 @@ func RandomByteGenerator(size int) []byte {
 	return builtBytes
 }
 
-// BitFlipper will flip the byteIndex bit in argument inByte, where index 0 is the MSB
-func BitFlipper(byteIndex int, inByte byte) byte {
-	if byteIndex < 0 || byteIndex > 7 {
-		fmt.Println("Error: BitFlipper() requires a byteIndex in range of 0-7")
-	}
-
-	for byteIndex < 0 || byteIndex > 7 {
-		fmt.Printf("Enter byteIndex in range of 0-7: ")
-		fmt.Scan(&byteIndex)
-	}
-
-	// Could also use inByte |= (1 << position) with a bit different logic...
-	var byteWithFlippedBit byte
-	switch byteIndex {
-	case 0: //! Flip MSB
-		byteWithFlippedBit = inByte ^ 128
-	case 1:
-		byteWithFlippedBit = inByte ^ 64
-	case 2:
-		byteWithFlippedBit = inByte ^ 32
-	case 3:
-		byteWithFlippedBit = inByte ^ 16
-	case 4:
-		byteWithFlippedBit = inByte ^ 8
-	case 5:
-		byteWithFlippedBit = inByte ^ 4
-	case 6:
-		byteWithFlippedBit = inByte ^ 2
-	case 7: //! LSB
-		byteWithFlippedBit = inByte ^ 1
-	}
-
-	return byteWithFlippedBit
-}
-
-// LeftByteShift shifts inByte by the amount of the  argument shiftBy
-func LeftByteShift(shiftBy int, inByte byte) byte {
-	shiftedLeftByte := inByte << shiftBy
-	return shiftedLeftByte
-}
-
-// RightByteShift shifts inByte by the amount of the  argument shiftBy
-func RightByteShift(shiftBy int, inByte byte) byte {
-	shiftedRightByte := inByte >> shiftBy
-	return shiftedRightByte
-}
-
-// RandomBitFlip will flip just 1 bit in a random Byte
-func RandomBitFlip(inData []byte) []byte {
-	randByteIndex := mrand.Intn(len(inData) - 1)
-	tempRandByte := inData[randByteIndex]
-	tempRandByte = BitFlipper(mrand.Intn(7), tempRandByte)
-	inData[randByteIndex] = tempRandByte
-
-	return inData
-}
-
-// RandomBitFlips flips a numBitsToFlip number of bits randomly
-func RandomBitFlips(inData []byte, numBitsToFlip int) []byte {
-	for i := 0; i < numBitsToFlip; i++ {
-		randByteIndex := mrand.Intn(len(inData) - 1)
-		tempRandByte := inData[randByteIndex]
-		tempRandByte = BitFlipper(mrand.Intn(7), tempRandByte)
-		inData[randByteIndex] = tempRandByte
-	}
-	return inData
-}
-
 // RandomByteFileGenerator will create a file with random bytes
 func RandomByteFileGenerator(size int, outFileName string) {
 	builtBytes := make([]byte, size)
@@ -192,18 +123,6 @@ func RandomByteFileGenerator(size int, outFileName string) {
 						less then length of random bytes to write`
 		fmt.Println(errorMessage)
 	}
-}
-
-// ByteFlipper flips all the bytes in the passed argument byte
-func ByteFlipper(byteToFlip byte) byte {
-	flippedBytes := ^byteToFlip
-	return flippedBytes
-}
-
-// ReverseByte reverses all the bits on the passed byteToReverse byte
-func ReverseByte(byteToReverse byte) byte {
-	reversedByte := bits.Reverse(uint(byteToReverse))
-	return byte(reversedByte)
 }
 
 // FileMutator is the start of a more complex function that will use different
@@ -252,8 +171,6 @@ func main() {
 
 	//Testing Flags + Parameters
 	//fmt.Println("flag.Args()[0]:", flag.Args()[0])
-	fmt.Println("flag.Args:", flag.Args())
-	fmt.Println("Args:", os.Args)
 
 	SetupLogger()
 
@@ -292,7 +209,6 @@ func main() {
 			fmt.Println("Error reading file: ", corpusName)
 			os.Exit(-1)
 		}
-		//fmt.Println(string(corpusFileBytes))
 	}
 
 	payloadSize := 0
@@ -318,17 +234,13 @@ func main() {
 		payloadSize = mrand.Intn(0x1fffffe8)
 	}
 
+	//Testing payload generation
 	if DEBUG {
 		fmt.Println("Randomly generated payload size: ", payloadSize)
 		payloadSize = 2000
 		fmt.Println("Debugging size: ", payloadSize)
-	}
-	//! Could make an additional arg to be passed depends on what is beingtotalNum
-	payload := string(RandomByteGenerator(payloadSize))
-	payload2 := string(RandomBitFlip(RandomByteGenerator(payloadSize)))
-
-	//Testing payload generation
-	if DEBUG {
+		payload := string(RandomByteGenerator(payloadSize))
+		payload2 := string(RandomBitFlip(RandomByteGenerator(payloadSize)))
 		fmt.Println("\nPayload_2: ", payload2)
 		fmt.Println("\nPayload: ", payload)
 	}
@@ -368,7 +280,8 @@ func main() {
 
 		log.Printf("-=-=####### Possible Crash:")
 		log.Printf("-=-=####### \tErr: %s", err.Error())
-		log.Printf("-=-=#######\t\tPayload @ Possible Crash:\n \"%s\"", payload)
+		//log.Printf("-=-=#######\t\tPayload @ Possible Crash:\n \"%s\"", payload)
+		//TODO: Create a dir Possible_Crashes, copy payload/payload file to it
 	}
 
 	fmt.Printf("Output of program: %s\n", string(output))
